removed projects, refactored structs and data cleaning
