+++ happ/zomes/projects/src/data.rs
@@ -1,0 +1,89 @@
+use hdk::prelude::*;
+use holo_hash::{HeaderHashB64, EntryHashB64};
+
+// Actual blob of data stored on the DHT
+#[hdk_entry(id = "thing")]
+#[derive(Clone)]
+#[serde(rename_all = "camelCase")]
+pub struct Thing {
+  pub data: String,
+}
+
+// returned after successful write to DHT
+#[derive(Serialize, Deserialize, Debug)]
+pub struct AddOutput {
+  pub header_hash: HeaderHashB64,
+  pub entry_hash: EntryHashB64,
+}
+
+// Passed in from UI
+#[derive(Serialize, Deserialize, Debug)]
+pub struct ThingInput {
+  pub path: String,
+  pub data: String,
+}
+
+// Sent back to UI
+#[derive(Clone, Serialize, Deserialize, Debug, Default, PartialEq)]
+pub struct Content {
+    pub name: String,
+    pub data: String,
+}
+
+#[derive(Clone, Serialize, Deserialize, Debug, Default)]
+pub struct Tree<T> 
+where
+    T: PartialEq
+{
+    pub tree: Vec<Node<T>>,
+}
+
+impl<T> Tree<T>
+where
+    T: PartialEq
+{
+    // create a new tree with a root node at index 0
+    pub fn new(root: T) -> Self {
+        Self {
+            tree: vec![Node::new(0, None, root)]
+        }
+    }
+
+    // inserts value into parent, return index of new node or 0 if parent doesn't exist
+    pub fn insert(&mut self, parent: usize, val: T) -> usize {
+        let idx = self.tree.len();
+        match self.tree.get_mut(parent) {
+            None => 0,
+            Some(node) => {
+                node.children.push(idx);
+                self.tree.push(Node::new(idx, Some(parent), val));
+                idx
+            }
+        }
+    }
+}
+
+#[derive(Clone, Serialize, Deserialize, Debug, Default)]
+pub struct Node<T>
+where
+    T: PartialEq
+{
+    pub idx: usize,
+    pub val: T,
+    pub parent: Option<usize>,
+    pub children: Vec<usize>,
+}
+
+impl<T> Node<T>
+where
+    T: PartialEq
+{
+    pub fn new(idx: usize, parent: Option<usize>, val: T) -> Self {
+        Self {
+            idx,
+            val,
+            parent,
+            children: vec![],
+        }
+    }
+}
