+++ happ/zomes/projects/src/lib.rs
@@ -1,47 +1,17 @@
-use std::collections::BTreeMap;
+mod test;
+mod data;
+mod tree_clean;
 
 use hdk::prelude::*;
 use holo_hash::{EntryHashB64, HeaderHashB64};
 use tracing::{info};
-mod project;
-mod test;
-
-use project::Project;
+use data::*;
+use tree_clean::{mark_tree, reindex_tree, prune_tree};
 
 entry_defs![
   PathEntry::entry_def(),
-  Thing::entry_def(),
-  Project::entry_def()
+  Thing::entry_def()
 ];
-
-// Actual blob of data stored on the DHT
-#[hdk_entry(id = "thing")]
-#[derive(Clone)]
-#[serde(rename_all = "camelCase")]
-pub struct Thing {
-  pub data: String,
-}
-
-// returned after successful write to DHT
-#[derive(Serialize, Deserialize, Debug)]
-pub struct AddOutput {
-  header_hash: HeaderHashB64,
-  entry_hash: EntryHashB64,
-}
-
-// Passed in from UI
-#[derive(Serialize, Deserialize, Debug)]
-pub struct ThingInput {
-  path: String,
-  data: String,
-}
-
-// Sent back to UI
-#[derive(Clone, Serialize, Deserialize, Debug, Default, PartialEq)]
-pub struct Content {
-    name: String,
-    data: String,
-}
 
 #[hdk_extern]
 pub fn put_thing(input: ThingInput) -> ExternResult<AddOutput> {
@@ -100,101 +70,6 @@ fn build_tree(tree: &mut Tree<Content>, node: usize, path: Path) -> ExternResult
   Ok(())
 }
 
-// Dead node marker
-fn mark_tree_for_delete(tree: &mut Tree<Content>, to_delete: &mut Vec<bool>) -> ExternResult<()>{
-  for node in tree.tree.clone().into_iter().rev() {
-    if node.children.len() == 0 && node.val.data.eq("") {
-      to_delete[node.idx] = true;
-    }
-
-    let mut are_children_deleted: bool =  true;
-    for child in node.children.iter() {
-      are_children_deleted = to_delete[*child] && are_children_deleted;
-    }
-
-    if are_children_deleted && node.val.data.eq("") && node.val.name != "root" {
-      to_delete[node.idx] = true;
-    }
-  };
-  Ok(())
-}
-
-// removed marked nodes
-fn prune_tree(tree: &mut Tree<Content>, pruned_tree: &mut Tree<Content>, to_delete: &mut Vec<bool>) -> ExternResult<()> {
-
-   // convert Vec<bool> to Vec<usize>
-   let mut idx_to_delete: Vec<usize> = vec![];
-   for (i, node) in to_delete.iter().enumerate() {
-     if *node {
-       idx_to_delete.push(i);
-     }
-   }
- 
-   // move non-deleted nodes to new tree
-   // update the children of those nodes 
-   for (idx, node) in to_delete.iter().enumerate() {
-     // if current node to delete is false, move it to new tree
-     if !*node {
-       let mut node: Node<Content> = tree.tree[idx].clone();
-       node.children.retain(|idx| !idx_to_delete.contains(idx));
-       pruned_tree.tree.push(node);
-     }
-   }
-  Ok(())
-}
-
-// update idx, parents and children
-fn reindex_tree(pruned_tree: Tree<Content>, reindexed_tree: &mut Tree<Content>) -> ExternResult<()> {
-  info!("Reindexing tree...");
-  // K: old idx of node V: new idx of node
-  // This map is populated as nodes get reindexed.
-  // This map get used to update the children Vec of nodes.
-  let mut idx_map: BTreeMap<usize, usize> = BTreeMap::new();
-
-  let rev_index: Vec<usize> = (0..pruned_tree.tree.len()).rev().collect();
-  let it = rev_index.iter().zip(pruned_tree.tree.iter().rev());
-  // crawl through nodes backwards and get map of old and new idxs
-  for (_, (i, node)) in it.enumerate() {
-    // nodes idx and location in tree Vec are mismatched. Condition to reindex.
-    if *i != node.idx {
-      // insert map of index change
-      idx_map.insert(node.idx, *i);
-    }
-  }
-
-  // for each node, update the parent and children from mapped idxs
-  // push onto reindexed_tree
-  for node in pruned_tree.tree.iter() {
-    let mut reindexed_node = node.clone();
-
-    for entry in idx_map.iter() {
-
-      // update idx
-      if node.idx == *entry.0 {
-        reindexed_node.idx = *entry.1;
-      }
-
-      // update children
-      for child in node.children.iter() {
-        if child == entry.0 {
-          reindexed_node.children.retain(|c| c != entry.0);
-          reindexed_node.children.push(*entry.1);
-        }
-      }
-
-      // update parent
-      if node.parent == Some(*entry.0) {
-        reindexed_node.parent = Some(*entry.1);
-      }
-    }
-    reindexed_tree.tree.push(reindexed_node);
-  }
-  
-  reindexed_tree.tree.sort_by(|a, b| a.idx.cmp(&b.idx));
-  info!("Reindexing tree complete.");
-  Ok(())
-}
-
 #[hdk_extern]
 pub fn get_thing(path_str: String) -> ExternResult<Option<Tree<Content>>> {
   let root_path = Path::from(path_str.clone());
@@ -211,7 +86,7 @@ pub fn get_thing(path_str: String) -> ExternResult<Option<Tree<Content>>> {
   build_tree(&mut tree, 0, root_path)?;
  
   let mut to_delete = vec![false; tree.tree.len()];
-  mark_tree_for_delete(&mut tree, &mut to_delete)?;
+  mark_tree(&mut tree, &mut to_delete)?;
   
   let mut pruned_tree: Tree<Content> = Tree { tree: vec![] };
   prune_tree(&mut tree, &mut pruned_tree, &mut to_delete)?;
@@ -267,60 +142,7 @@ pub fn try_get_element(entry_hash: EntryHash, get_options: GetOptions) -> Extern
   }
 }
 
-#[derive(Clone, Serialize, Deserialize, Debug, Default)]
-pub struct Tree<T> 
-where
-    T: PartialEq
-{
-    pub tree: Vec<Node<T>>,
-}
 
-impl<T> Tree<T>
-where
-    T: PartialEq
-{
-    // create a new tree with a root node at index 0
-    pub fn new(root: T) -> Self {
-        Self {
-            tree: vec![Node::new(0, None, root)]
-        }
-    }
 
-    // inserts value into parent, return index of new node or 0 if parent doesn't exist
-    pub fn insert(&mut self, parent: usize, val: T) -> usize {
-        let idx = self.tree.len();
-        match self.tree.get_mut(parent) {
-            None => 0,
-            Some(node) => {
-                node.children.push(idx);
-                self.tree.push(Node::new(idx, Some(parent), val));
-                idx
-            }
-        }
-    }
-}
 
-#[derive(Clone, Serialize, Deserialize, Debug, Default)]
-pub struct Node<T>
-where
-    T: PartialEq
-{
-    idx: usize,
-    val: T,
-    parent: Option<usize>,
-    children: Vec<usize>,
-}
 
-impl<T> Node<T>
-where
-    T: PartialEq
-{
-    fn new(idx: usize, parent: Option<usize>, val: T) -> Self {
-        Self {
-            idx,
-            val,
-            parent,
-            children: vec![],
-        }
-    }
-}
