--- happ/zomes/projects/src/project/mod.rs
@@ -1,5 +1,0 @@
-mod handlers;
-mod entry;
-
-pub use handlers::*;
-pub use entry::*;
